requirements.txt
flask
gunicorn
python-dotenv
requests

google-generativeai
openai

pandas
numpy
